% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_na}
\alias{convert_to_na}
\title{Convert values that we consider valid missing values to the NULL object}
\usage{
convert_to_na(col_name)
}
\arguments{
\item{col_name}{(string or NULL object)}
}
\value{
col_name (string or NULL object)
}
\description{
Convert values that we consider valid missing values to the NULL object
}
