% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_chart.R
\name{create_demo_chart}
\alias{create_demo_chart}
\title{Function to visualize demographic disparity scores on a lollipop chart}
\usage{
create_demo_chart(
  demo_df,
  group = "total",
  save_chart = FALSE,
  file_path = "dem_disparity_chart.png"
)
}
\arguments{
\item{demo_df}{(data.frame) - the demographic disparity score dataframe
returned by the SEDT API}

\item{group}{(character) - One of "total", "poverty", or "under18". Default
is "total" This variable indicates the demographic group for which the chart
to show.}

\item{save_chart}{(logical) - Default set to FALSE. Whether to save the chart
or not.}

\item{file_path}{(character) - Default set to "dem_disparity_chart.png".
A file path of where to save the file. This should include a data type
suffix. EX: "results/visuals/dem_disparity_chart.png"}
}
\value{
plot (ggplot object)  - The ggplot object that was created.
}
\description{
Function to visualize demographic disparity scores on a lollipop chart
}
\details{
Please use the following citation for data obtained from the API,
replacing the version number with the current version:

Stern, Alena, Gabe Morrison, Sonia Torres Rodríguez, Ajjit Narayanan, and Graham MacDonald. 2024.
“Spatial Equity Data Tool API” (Version X.x.x). Washington, DC: Urban Institute.
https://ui-research.github.io/sedt_documentation/. Data originally sourced from various sources,
analyzed at the Urban Institute and made available under the ODC Attribution License.
}
