% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Function to create a choropleth map to visualize the geographic disparity scores}
\usage{
create_map(
  geo_df,
  col_to_plot = "diff_pop",
  save_map = FALSE,
  interactive = TRUE,
  file_path = "bias_map"
)
}
\arguments{
\item{geo_df}{(sf dataframe) - a spatial dataframe containing the geographic
disparity scores outputted from the Spatial Equity Data Tool.}

\item{col_to_plot}{(string) - The column in the geo_df to plot in
choropleth map.}

\item{save_map}{(logical) - Default set to FALSE. Whether to save the chart
or not. Note that if interactive is set to true, the map will save as a
.html. Otherwise, the map will save as a .png.}

\item{interactive}{(logical) - Default set to TRUE. Whether the map should
be interactive or not.}

\item{file_path}{(character) - Default set to "bias_map". A file path of where
to save the file. This should not include a file type as that is controlled
by the interactive variable. An example file-path would be,
"visuals/interactives/disparity_map".}
}
\value{
bias_map (tmap map) - the choropleth map created
}
\description{
Function to create a choropleth map to visualize the geographic disparity scores
}
\details{
Please use the following citation for data obtained from the API,
replacing the version number with the current version:

Stern, Alena, Gabe Morrison, Sonia Torres Rodríguez, Ajjit Narayanan, and Graham MacDonald. 2024.
“Spatial Equity Data Tool API” (Version X.x.x). Washington, DC: Urban Institute.
https://ui-research.github.io/sedt_documentation/. Data originally sourced from various sources,
analyzed at the Urban Institute and made available under the ODC Attribution License.
}
