% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{geo_bias_map}
\alias{geo_bias_map}
\alias{plot_geo_bias_map}
\title{Plot a map of geo bias data with \code{{tmap}} or \code{{ggplot2}}}
\usage{
plot_geo_bias_map(
  data,
  fill_col = "diff_pop",
  fill_label = "Disparity Score",
  fill_palette = "RdBu",
  fill_scale = ggplot2::scale_fill_distiller(type = "div", palette = fill_palette, labels
    = scales::label_percent(scale = 1), direction = 1),
  plot_theme = ggplot2::theme_void()
)
}
\arguments{
\item{fill_col}{String with column name to map to fill.}

\item{fill_label}{Label to use for fill column. Passed as title for
\code{\link[tmap:tm_polygons]{tmap::tm_fill()}} or fill label for \code{\link[ggplot2:labs]{ggplot2::labs()}}.}

\item{fill_palette}{String with palette name.}

\item{fill_scale}{ggplot2 scale function to use with map.}

\item{plot_theme}{ggplot2 theme to use with map.}
}
\description{
\code{\link[=tm_plot_geo_bias_map]{tm_plot_geo_bias_map()}} and \code{\link[=plot_geo_bias_map]{plot_geo_bias_map()}} are internal plotting
functions called by the more generic \code{\link[=create_map]{create_map()}} function.
}
\keyword{internal}
