% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_api_response}
\alias{get_api_response}
\title{Get API Response}
\usage{
get_api_response(param_list)
}
\arguments{
\item{param_list}{(string) Required: Includes the parameters included in the
call_upload_user_files function}
}
\value{
r (string) Returns the API error status code response
}
\description{
Calls \code{\link[=call_upload_user_files]{call_upload_user_files()}} for a given file path, lat col,
and lon col, parses and returns API response
}
