% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Title}
\usage{
get_status(file_id)
}
\arguments{
\item{file_id}{(string) id string corresponding to a POST request that was previously submitted to
the /upload-user-file endpoint}
}
\value{
r_list (list) An R list of a JSON object with information about the submission to the SEDT
}
\description{
Title
}
