% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_col_type}
\alias{is_valid_col_type}
\title{Is Valid Col Type}
\usage{
is_valid_col_type(col_name)
}
\arguments{
\item{col_name}{(string or NULL object)}
}
\value{
(boolean)
}
\description{
Tests if the value of the baseline columns or demographic columns
dictionaries has a valid data type
}
