% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sedt-citation}
\alias{sedt-citation}
\title{Citation for Spatial Equity Data Tool API Data}
\description{
Citation for Spatial Equity Data Tool API Data
}
\details{
Please use the following citation for data obtained from the API:

Stern, Alena, Gabe Morrison, Sonia Torres Rodríguez, Ajjit Narayanan, and
Graham MacDonald. 2024. “Spatial Equity Data Tool API” (Version
0.0.0.4000). Washington, DC: Urban Institute.
\url{https://ui-research.github.io/sedt_documentation/}. Data originally sourced
from various sources, analyzed at the Urban Institute and made available
under the ODC Attribution License.
}
