% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{tm_plot_geo_bias_map}
\alias{tm_plot_geo_bias_map}
\title{\code{\link[=tm_plot_geo_bias_map]{tm_plot_geo_bias_map()}} uses \code{\link[tmap:tm_shape]{tmap::tm_shape()}} and \code{\link[tmap:tm_polygons]{tmap::tm_fill()}} to map
the geo bias data returned by \code{\link[=call_sedt_api]{call_sedt_api()}}.}
\usage{
tm_plot_geo_bias_map(
  data,
  fill_col = "diff_pop",
  fill_label = "Disparity Score",
  basemap_server = "CartoDB.PositronNoLabels",
  border_lwd = 0.25,
  fill_palette = "RdBu",
  title = NULL,
  legend.outside = TRUE,
  attr.outside = TRUE,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{title}{Global title of the map. For small multiples, multiple titles can be specified. The title is drawn inside the map. Alternatively, use \code{panel.labels} to print the map as a panel, with the title inside the panel header (especially useful for small multiples). Another alternative is the \code{main.title} which prints a title above the map. Titles for the legend items are specified at the layer functions (e.g. \code{\link[tmap]{tm_fill}}).}

\item{legend.outside}{Logical that determines whether the legend is plot outside of the map/facets. Especially useful when using facets that have a common legend (i.e. with \code{free.scales=FALSE}).}

\item{attr.outside}{Logical that determines whether the attributes are plot outside of the map/facets.}

\item{...}{other arguments from \code{tm_layout}}
}
\description{
\code{\link[=tm_plot_geo_bias_map]{tm_plot_geo_bias_map()}} uses \code{\link[tmap:tm_shape]{tmap::tm_shape()}} and \code{\link[tmap:tm_polygons]{tmap::tm_fill()}} to map
the geo bias data returned by \code{\link[=call_sedt_api]{call_sedt_api()}}.
}
\keyword{internal}
