% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_chart.R
\name{create_demo_chart}
\alias{create_demo_chart}
\title{Function to visualize demographic disparity scores on a lollipop chart}
\usage{
create_demo_chart(
  demo_df,
  group = "total",
  save_chart = TRUE,
  file_path = "dem_disparity_chart.png"
)
}
\arguments{
\item{demo_df}{(data.frame) - the demographic disparity score dataframe
returned by the SEDT API}

\item{group}{(character) - One of "total", "poverty", or "under18". Default
is "total" This variable indicates the demographic group for which the chart
to show.}

\item{save_chart}{(logical) - Default set to TRUE. Whether to save the chart
or not.}

\item{file_path}{(character) - Default set to "dem_disparity_chart.png".
A file path of where to save the file. This should include a data type
suffix. EX: "results/visuals/dem_disparity_chart.png"}
}
\value{
plot (ggplot object)  - The ggplot object that was created.
}
\description{
Function to visualize demographic disparity scores on a lollipop chart
}
