% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create a chloropleth map to visualize the geographic disparity scores}
\usage{
create_map(
  geo_df,
  col_to_plot = "diff_pop",
  ...,
  pkg = "tmap",
  interactive = TRUE,
  save_map = FALSE,
  save_args = list(),
  file_path = "bias_map",
  file_ext = "png"
)
}
\arguments{
\item{geo_df}{(sf dataframe) - a spatial dataframe containing the geographic
disparity scores outputted from the Spatial Equity Data Tool.}

\item{col_to_plot}{(string) - The column in the geo_df to plot in
choropleth map.}

\item{pkg}{(string) Package to use when creating the map. One of "tmap"
(default) or "ggplot2".}

\item{interactive}{(logical) - Default set to TRUE. Whether the map should be
interactive or not.}

\item{save_map}{(logical) - Default set to FALSE. Whether to save the chart
or not. Note that if \code{interactive = TRUE} and \code{pkg = "tmap"}, the map will
save as a .html. Otherwise, the map will save as a .png.}

\item{file_path}{(character) - Default set to "bias_map". A file path of where
to save the file. This should not include a file type as that is controlled
by the interactive variable. An example file-path would be,
"visuals/interactives/disparity_map".}
}
\value{
bias_map (tmap map or ggplot2 plot) - the choropleth map created
}
\description{
\code{\link[=create_map]{create_map()}} uses \code{{tmap}} and \code{{urbnthemes}} to create a chloropleth map
to visualize the geographic disparity scores returned with the response
object from \code{\link[=call_sedt_api]{call_sedt_api()}}.
}
\details{
Please use the following citation for data obtained from the API:

Stern, Alena, Gabe Morrison, Sonia Torres Rodríguez, Ajjit Narayanan, and
Graham MacDonald. 2024. “Spatial Equity Data Tool API” (Version
0.0.0.4000). Washington, DC: Urban Institute.
\url{https://ui-research.github.io/sedt_documentation/}. Data originally sourced
from various sources, analyzed at the Urban Institute and made available
under the ODC Attribution License.
}
