% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_data.R
\name{get_output_data}
\alias{get_output_data}
\title{Function to call get-equity-file api endpoint}
\usage{
get_output_data(file_id)
}
\arguments{
\item{file_id}{(string) - the id associated with the api function call. This
should be returned from call_upload_user_files.R}
}
\value{
list of six elements:
status_code (integer) - a status code showing whether the API call was successful
file_exists (Boolean) - whether or not the equity status data exists.
file_id (string) - the inputed file_id
geo_bias_data (sf dataframe) - the geographic bias dataset outputted from
SEDT API. NA if file_exists is FALSE.
demo_bias_data (dataframe) - the demographic bias dataset outputted from the
SEDT API. NA if file_exists is FALSE.
full_api_results (list) - the entirety of the JSON response object from the
get-equity-file api endpoint converted to an R list. NA if file_exists is FALSE.
}
\description{
Function to call get-equity-file api endpoint
}
