% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_resource.R
\name{obj_as_point}
\alias{obj_as_point}
\title{Convert a sfg, sfc, or sf object to use POINT geometry}
\usage{
obj_as_point(
  x,
  split_multi_point = FALSE,
  placement = c("centroid", "surface"),
  allow_type = c("POINT", "MULTIPOINT", "LINESTRING", "MULTILINESTRING", "POLYGON",
    "MULTIPOLYGON"),
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{A sfg, sfc, or sf object to convert to POINT geometry.}

\item{placement}{If "centroid" (default) convert geometry to POINT using
\code{\link[sf:geos_unary]{sf::st_centroid()}}. If "surface", use \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}.}

\item{allow_what}{Allowed types of geometry. Defaults to \code{c("POINT", "MULTIPOINT", "LINESTRING", "MULTILINESTRING", "POLYGON", "MULTIPOLYGON")}.}
}
\description{
Convert a sfg, sfc, or sf object to use POINT geometry
}
\keyword{internal}
