% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedtR-package.R
\docType{package}
\name{sedtR-package}
\alias{sedtR}
\alias{sedtR-package}
\title{sedtR: sedtR allows programmers to call the Urban Institute's Spatial Equity Data Tool API}
\description{
The Urban Institute has developed the Spatial Equity Data Tool to analyze the equitability of resources across cities, counties, states, and the entirety of the United States. This package allows users to call the Spatial Equity Data Tool API programmatically using R code.
}
\author{
\strong{Maintainer}: Gabriel Morrison \email{gmorrison@urban.org}

Authors:
\itemize{
  \item Alena Stern \email{astern@urban.org}
}

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
}

}
\keyword{internal}
