% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_resource.R
\name{weight_resource_by_area}
\alias{weight_resource_by_area}
\alias{weight_resource_by_admin_geo}
\title{Weight a resource based on intersection with administrative areas}
\usage{
weight_resource_by_area(
  resource,
  area,
  weight = c("length", "area"),
  placement = "surface",
  ...,
  call = caller_env()
)

weight_resource_by_admin_geo(
  resource,
  state = NULL,
  county = NULL,
  geo = "city",
  admin_geo = "tract",
  weight = c("length", "area"),
  placement = "surface",
  year = getOption("sedtR.year", 2021),
  ...,
  call = caller_env()
)
}
\arguments{
\item{resource}{A \code{sf} object to modify.}

\item{area}{A \code{sf} or \code{sfc} object to intersect with resource.}

\item{weight}{Type of function to use in calculating the weight value. If
"length", use \code{\link[sf:geos_measures]{sf::st_length()}} and, if "area", use
\code{\link[sf:geos_measures]{sf::st_area()}} to calculate weight values based on the geometry. \code{weight}
can also be any other function that supports the geometry column as the
only input parameter. If \code{weight} is a function, the results are coerced to
a numeric vector (dropping the units class if available).}

\item{placement}{Type of function to use in processing the returned geometry.
"surface" (default) uses \code{\link[sf:geos_unary]{sf::st_point_on_surface()}}, "centroid" uses
\code{\link[sf:geos_unary]{sf::st_centroid()}}, or "none" returns geometry with the same type as the
input resource modified only by the intersection with the specified
administrative geography.}

\item{...}{Additional parameters passed to tigris function specified by
\code{admin_geo}.}

\item{geo}{One of "city", "county", "state", or "national". Not yet
implemented but this parameter should be used for validating the
\code{admin_geo} value.}

\item{admin_geo}{Type of geometry to use for intersecting area. One of
"tract" (default), "county", or "state" specifying the corresponding tigris
function (\code{tigris::tracts}, \code{tigris::counties}, or \code{tigris::states}).}

\item{year}{Year of data passed to tigris functions. Uses global
\code{"sedtR.year"} option or 2021 if option is not set.}
}
\value{
A modified sf object supplied to \code{resource} where features are
intersected with \code{area} and optionally transformed to use POINT geometry.
}
\description{
\code{\link[=weight_resource_by_area]{weight_resource_by_area()}} uses \code{\link[sf:geos_binary_ops]{sf::st_intersection()}} to combine a
\code{resource} sf object with \code{area} geometry to derive a new weight column based
on the length or area of the geometry created by the intersection.
\code{\link[=weight_resource_by_admin_geo]{weight_resource_by_admin_geo()}} is a variation that uses the \code{{tigris}}
package to download administrative areas from the U.S. Census Bureau to use
as the \code{area} input.
}
\examples{
\dontrun{
if (interactive()) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))

  roads <- tigris::roads(state = "NC", county = nc$NAME[1:5])

  weight_resource_by_area(
    roads,
    nc[1:5, ],
    weight = "length"
  )
}
}
}
