% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_resource.R
\name{weight_resource_by_area}
\alias{weight_resource_by_area}
\alias{weight_resource_by_admin_geo}
\title{Weight a resource based on intersection with administrative areas}
\usage{
weight_resource_by_area(
  resource,
  area,
  weight = c("length", "area"),
  placement = c("centroid", "surface", "none"),
  ...
)

weight_resource_by_admin_geo(
  resource,
  state = NULL,
  county = NULL,
  geo = "city",
  admin_geo = "tract",
  weight = c("length", "area"),
  placement = c("centroid", "surface", "none"),
  ...,
  call = caller_env()
)
}
\arguments{
\item{resource}{A sf object to modify.}

\item{area}{A sf or sfc object to intersect with resource.}

\item{weight}{If "length", use \code{\link[sf:geos_measures]{sf::st_length()}} to assign weight. If "area",
use \code{\link[sf:geos_measures]{sf::st_area()}} to assign weight.}

\item{placement}{One of "centroid", "surface", or "none".}

\item{...}{arguments to be passed to internal function \code{load_tiger}, which is not exported. See Additional Arguments.}

\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation. When \code{NULL} and combined with \code{cb = TRUE}, a national dataset of Census tracts will be returned for years 2019
and later.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties. Can
also be a county name or vector of names.}

\item{geo}{One of "city", "county", "state", or "national".}

\item{admin_geo}{One of "tract", "county", or "state".}
}
\description{
\code{\link[=weight_resource_by_area]{weight_resource_by_area()}} uses \code{\link[sf:geos_binary_ops]{sf::st_intersection()}} to combine
\code{resource} with \code{area} to derive a new weight column based on the length or
area of the intersected geometry. \code{\link[=weight_resource_by_admin_geo]{weight_resource_by_admin_geo()}} is a
variation that uses the \code{{tigris}} package to download administrative areas
from the U.S. Census Bureau to use as the \code{area} input.
}
\keyword{internal}
